Add files via upload
This system aims to provide a solution that assist in a Barista (Server) servicing multiple Customers (Clients).The purpose of the Barista will be to manage multiple orders from different Customers whilst minimising waste by repurposing abandoned orders to new orders if the requirements meet.  The purpose of the Customer is to make orders to the Barista. At any point in time the Customer can make a request to the Barista to find out their order status. The barista is responsible for returning the correct order status to the customer as required. Throughout this report we will identify the Server as Barista and the Client as Customer.